CHROMIUM: ASoC: img: Add a card driver for Concerto-based boards
Add an audio card driver for boards based on the Concerto module.
This card driver is capable of driving all CPU DAIs on the Pistachio
SoC with optional codec DAIs on the I2S playback and capture paths.

Note that this driver is currently very simple.  It does not yet support:
 - audio PLL adjustment,
 - handling of conflicts between I2S in/out MCLK rate selection,
 - I2S in CPU clock master support,
 - auxiliary devices (e.g. amplifiers).

BUG=chrome-os-partner:37173
TEST=I2S-out playback and parallel-out playback on Kennet.

Change-Id: Idcc1c57e6dc10b0f7bdd5d16517530a9540975db
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/255736
Reviewed-by: Kevin Cernekee <cernekee@chromium.org>