UPSTREAM: usb: dwc2: pci: Add device mode to the dwc2-pci driver
The pci driver now registers a platform driver, like in dwc3, and lets
its probe function do all the initialization. This allows it to
account for changes to the platform driver that were not added to the
pci driver. Also future changes to the probe function don't have to be
duplicated. This also has the effect of adding device and DRD mode to
the pci driver. Tested on the Synopsys HAPS PCIe platform.

BUG=chrome-os-partner:41748
TEST=w/ series USB is happy, even with s2r

Signed-off-by: John Youn <johnyoun@synopsys.com>
Signed-off-by: Felipe Balbi <balbi@ti.com>
Signed-off-by: Doug Anderson <dianders@chromium.org>
(cherry picked from commit 9024c495f35be735a917571406fab30a789c27d1)
Change-Id: Id1f6b5cf22b2444029a092b86106e2179439cf69
Reviewed-on: https://chromium-review.googlesource.com/281318
Reviewed-by: yunzhi li <lyz@rock-chips.com>
Reviewed-by: Julius Werner <jwerner@chromium.org>