CHROMIUM: ASoC: img: Add support for I2S clock loopback
Pistachio can be configured to loopback the i2s-out LRCLK/BCLK
signals to the i2s-in block, which can allow Pistachio to act
as the clock master while doing I2S input even though the i2s-in
block only supports being the clock slave.  Add the device-tree
property "img,i2s-clock-loopback" (specified in the i2s-in CPU node)
which indicates that the internal clock loopback should be enabled.

BUG=chrome-os-partner:37173,chrome-os-partner:35348
TEST=Record with the Pistachio as the master on Kennet.

Change-Id: I43d71076c210a26cbb19da522b4c1400ad1d9d23
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/259360
Reviewed-by: Kevin Cernekee <cernekee@chromium.org>