CHROMIUM: iwl7000: system and firmware debug improvements
This is a combination of various patches we've developed to improve
system behaviour and firmware debugging. Notably, this allows us to
specify a set of host-based triggers (like beacon loss statistics)
that could be related to firmware issues in particular debugging
scenarios, and then dump the firmware state on these.

BUG=N/A
TEST=AVL test run

Change-Id: I3dda0fc6ca564e589d9eb4dca4d6beaffbf0b95c
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/256043
Reviewed-by: Paul Stewart <pstew@chromium.org>