Revert "FROMLIST: ath10k: fix quiet duration configuration"
This reverts commit a42dd4dd0e6235cbeb861dcf940e1108d91cc0da.

This change inverts the ath10k interface for thermal throttling.
Previously writing 0 to the sysfs attribute meant '0% throttling', with
this change it means '0% duty cycle'. The user space thermal daemon uses
the initial interface, now WiFi stops working when the daemon sets the
duty cycle to 100%.

The driver uses a kernel cooling_device to configure
the throttling, therefore the previous behavior was correct:

max_state: Maximum cooling state of the cooling device

https://www.kernel.org/doc/Documentation/thermal/sysfs-api.txt

BUG=chrome-os-partner:37507
TEST=Manual on whirlwind

Change-Id: I9c9a44551027d2e50d9a6ea0e180404ccf9697b8
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/256625
Reviewed-by: Srinivasa duvvuri <sduvvuri@chromium.org>
Reviewed-by: Bill Hegardt <bhegardt@chromium.org>
Reviewed-by: Toshi Kikuchi <toshik@chromium.org>