CHROMIUM: Add smart antenna APIs
This patch defines APIs used for smart antenna algorithm.
Has a full implementation os wmi ops used for smart antenna
configuration. There is a new module param which is used to
enable smart antenna logic on host and firmware.

Signed-off-by: Vasanthakumar Thiagarajan <vthiagar@codeaurora.org>

BUG=chrome-os-partner:37796
TEST=build/boot. Load ath10k_core with modparam set enable_smart_antenna=1,
bring up AP mode using hostapd and connect a station to this AP.
Device was able to connect to the internet and load multiple websites.
This testing is done on AP148.

Change-Id: I9b1178e6296d0bdbe5d8161562f5da3bce581e19
Reviewed-on: https://chromium-review.googlesource.com/254498
Reviewed-by: Toshi Kikuchi <toshik@chromium.org>
Tested-by: Peter Oh <poh@qca.qualcomm.com>
Commit-Queue: Toshi Kikuchi <toshik@chromium.org>