CHROMIUM: ARM: dts: Don't enable i2c3 on veyriants that don't use it
The i2c3 bus is only present on minnie and pinky.  On other boards it
is often no connected with no external pullups so is not really a
valid bus.  Let's avoid confusing anyone probing i2c busses (and save
a tiny amount of memory) by disabling i2c3 where it's not used.

BUG=chrome-os-partner:35216
TEST=i2cdetect -l | grep i2c-3 => nothing for boards except minnie

Change-Id: I563e72aa6b8bfde47377f943a09b7b7a17f4a36c
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/257072
Reviewed-by: Alexandru Stan <amstan@chromium.org>