FROMLIST: ath10k: Fix interpretation of cooling device state
Setting the sysfs attribute ends up configuring the duty cycle,
but the interface through which the attribute is exposed
(cooling_device) is for setting the throttle/cooling state. This
is confusing the user. Hence renaming the cooling device interfaces
for better readability.

BUG=chrome-os-partner:37767
TEST=build/boot on storm suceeded. Setup Storm board as AP using hostapd and
connected to the board using another device. Device was able to connect to
the internet and load multiple websites.

Cc: Matthias Kaehlcke <mka@google.com>
Signed-off-by: Rajkumar Manoharan <rmanohar@qti.qualcomm.com>

Change-Id: Id856972f632fba9348d66464bc46107bcc2281a3
Reviewed-on: https://chromium-review.googlesource.com/257410
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>
Tested-by: Matthias Kaehlcke <mka@chromium.org>