FROMLIST: ath10k: cache throttle state when device is down
Allow driver to cache the throttle state when the devie is not
yet started. Configure the cached throttle state while powering
up the device. Since thermal daemon is unaware of the up/down cycle,
it assumes that device is throttled.

BUG=chrome-os-partner:36990
TEST=build/boot on storm suceeded. Setup Storm board as AP using hostapd and
connected to the board using another device. Device was able to connect to
the internet and load multiple websites.

Reported-by: Matthias Kaehlcke <mka@google.com>
Signed-off-by: Rajkumar Manoharan <rmanohar@qti.qualcomm.com>

Change-Id: Ib228405d8dcf2ffff3bfbf41ad7461cb7fb84bad
Reviewed-on: https://chromium-review.googlesource.com/257413
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Toshi Kikuchi <toshik@chromium.org>
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>