CHROMIUM: ASoC: rt5677: Stop and restart DSP over suspend/resume
MCLK1 gets disabled at suspend and re-enabled at resume. Before
MCLK1 is re-enabled, if the DSP is already on (either the DSP was
left on during suspend, or the DSP is turned on early at resume),
i2c register read returns garbage and corrupts the regmap cache.

This patch stops the DSP before suspend and restarts it after
resume with a dalay to ensure MCLK is on while loading firmware.

BUG=chrome-os-partner:37540
TEST=Hotwording works after repeated suspend/resmue with or without
background audio playback.

Change-Id: I8a64a2770a76661175927ce75c2e1b54b2f31814
Signed-off-by: Ben Zhang <benzh@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/259291
Reviewed-by: Dylan Reid <dgreid@chromium.org>