CHROMIUM: rockchip: Add read valid training to dmc freq
This adds read valid training to dmc freq. Without it, certain frequencies could
run into issues with memory corruption. This increases the timeout for dmc freq
to account for the extra time performing data training. It also adds in more
refresh commands to have the same amount as in firmware since we're doing the
same training sequences.

BUG=chrome-os-partner:34608
TEST=boot speedy

Change-Id: I98f1f85f9d30ab1f2f44870e69f533a41463c095
Signed-off-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/257990
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Tested-by: Caesar Wang <wxt@rock-chips.com>