FROMLIST: ath10k: move driver state check before setting throttle
Since thermal daemon is unaware of the device state, it might
try to adjust the throttle state when the device is powered down.
So the driver caches the value and will configure it while
powering up the target. The cached value will be programed later
once the device is brought up. In such case, returning error
status is confusing and misleading the user application. Hence
moving the driver state check before sending wmi command to target.

BUG=chrome-os-partner:36990
TEST=on whirlwind:
  ifconfig wifi2 down
  echo 70 > /sys/class/net/wifi2/device/cooling_device/cur_state
    => no error message

Reported-by: Matthias Kaehlcke <mka@google.com>
Signed-off-by: Rajkumar Manoharan <rmanohar@qti.qualcomm.com>

Change-Id: I2b55765e9816f5d5b3ddd1af85075c3beacb65f5
Reviewed-on: https://chromium-review.googlesource.com/259093
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Toshi Kikuchi <toshik@chromium.org>
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>