ANDROID: seccomp: really fix syscall numbers for x86 and x86_64
This fixes a build break on x86/x86_64 introduced by f2caf97
"seccomp: fix syscall numbers for x86 and x86_64".

The original change looks like it was cherry-picked from a kernel
with sys_renameat2, but since this kernel (android-3.14) lacks an
implementation of that syscall, it is not correct to refer to it
in the x86 syscall tables, or vmlinux will fail to link.

Do what ARM does and define the syscall slot but have it wired
up to the sys_ni_syscall / compat_ni_syscall so we retain syscall
binary compatibility with newer kernels.

Signed-off-by: Alistair Strachan <alistair.strachan@imgtec.com>
(cherry picked from commit 35bab84a0e530d4ce42a3ce84c3976ce8017eac8)

Change-Id: Ibe34f8e19f1daeb0895255d3e0a4e24d319dce44
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/285958
Reviewed-by: Olof Johansson <olofj@chromium.org>