CHROMIUM: pinctrl: pistachio: Prepare the support for other devices
The Pistachio SoC contains two sets of pins: east side and west side.
The east side pins are associated to the MIPS core, controlling the
peripherals and six banks of GPIOs. The west side pins are associated
to the RPU core, for bluetooth and wireless devices. The west side
has access to two banks of GPIOs.

The register layout for the two pin control sets is identical, but
the number of pins and functions vary. Currently, the driver supports
only the east side.

This commit adds an infrastructure to support multiple sets of pins.

BUG=chrome-os-partner:35348
TEST=Tested on Pistachio BuB

Change-Id: Iddae1908fedd5fabb5621fadc34193e801f6f013
Signed-off-by: Ezequiel Garcia <ezequiel.garcia@imgtec.com>
Reviewed-on: https://chromium-review.googlesource.com/259990
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>