CHROMIUM: MIPS: dts: buranku: Add audio input support
Add support for the I2S input path on Buranku.  The SoC is the clock
master on the input path, so this also requires the I2S output path
to be enabled as well in order to use the clock loopback.  Note that
playback is not yet supported, though.

On rev1 boards the aux input is connected to I2S channel 0 and a DMIC
is connected to I2S channel 1.  On rev2 boards DMICs are connected to
I2S channels 0 and 1 while the aux input is connected to I2S channel 4.

Note that rev1 boards require a rework to connect the ADC and DMIC
clock lines to I2S_OUT_{LRCLK,BCLK} rather than I2S_IN_{LRCLK,BCLK}.

BUG=chrome-os-partner:37173
TEST=Record from aux in and DMIC on Buranku rev1.

Change-Id: I33fe6ed6cf1ec7f00a91329ab3dc2f5154ab3776
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/260063
Reviewed-by: Kevin Cernekee <cernekee@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>