CHROMIUM: pinctrl: pistachio: Use pin_base for the GPIO base
The driver currently sets the GPIO base as bank * 16. While this works
when only the system pinctrl set is supported, it won't work well when
the rpu set is introduced.

This commit changes the way gpio_chip and gpios_ranges base are specified,
by making them equal to the pin_base value.

BUG=chrome-os-partner:35348
TEST=Tested on Pistachio BuB

Change-Id: I934ae6d66901ef02b679637f6cef224332a29026
Signed-off-by: Ezequiel Garcia <ezequiel.garcia@imgtec.com>
Reviewed-on: https://chromium-review.googlesource.com/260410
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>