UPSTREAM: usb: dwc2: pci: Correctly compile dwc2-pci as a module or built-in
The dwc2-pci driver should be compiled as a module when configured to
do so. If the dwc2-pci is configured as a module but actually
built-in, it can cause build errors due to the fact that the
generic-phy will be allowed to compile as a module causing undefined
references.

BUG=chrome-os-partner:41748
TEST=w/ series USB is happy, even with s2r

Reported-by: Alan Cox <alan@lxorguk.ukuu.org.uk>
Signed-off-by: John Youn <johnyoun@synopsys.com>
Signed-off-by: Felipe Balbi <balbi@ti.com>
Signed-off-by: Doug Anderson <dianders@chromium.org>
(cherry picked from commit b48cb02de37976ad2ebef201ad40aa40156f6daa)
Change-Id: I67f7fe1bf6201781393fd2294f8bf040c5132138
Reviewed-on: https://chromium-review.googlesource.com/281320
Reviewed-by: yunzhi li <lyz@rock-chips.com>
Reviewed-by: Julius Werner <jwerner@chromium.org>