CHROMIUM: crypto: img-hash: Improve debug output
Currently each time a hash is processed by the hardware, a message
is printed to the console.  This is unnecessarily verbose, so
change this to only happen when debug is specifically enabled.

In addition, the successful probe of the hardware only results
in a console message when debug is enabled, but it would be
useful for this to happen as standard. Enable it even when debug
is not enabled.

TEST=Run on the pistachio BuB
BUG=chrome-os-partner:35348

Change-Id: Ie032bcbf0338947c316d823d0e61a65ce0d3967f
Signed-off-by: James Hartley <james.hartley@imgtec.com>
Reviewed-on: https://chromium-review.googlesource.com/261340
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>
Commit-Queue: Andrew Bresticker <abrestic@chromium.org>