CHROMIUM: brcmfmac: Use static superset of channels for wiphy bands
The driver was constructing a list of channels per wiphy band
by querying the device. This list is not what the hardware is
able to do as it is already filtered by the country setting in
the device. As user-space may change the country this would
require updating the channel list which is not recommended [1].
This patch introduces a superset of channels. The individual
channels are disabled appropriately by querying the device.

[1] http://mid.gmane.org/1426706320.3001.21.camel@sipsolutions.net
BUG=chrome-os-partner:38226
TEST=manual(iw reg command)
Signed-off-by: Arend van Spriel <arend@broadcom.com>
Signed-off-by: Daniel Kim <dekim@broadcom.com>

Change-Id: Id72a1f000e0001126adc066d59b49ab986403dd1
Reviewed-on: https://chromium-review.googlesource.com/261627
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Tested-by: mukesh agrawal <quiche@chromium.org>
Commit-Queue: mukesh agrawal <quiche@chromium.org>