CHROMIUM: spi: img-spfi: Control chipselect from GPIO
When the continue bit is set, the interrupt status we are polling
to identify if a transaction has finished can be sporadic.
Even though the transfer has finished the interrupt status will say:
"All done, nothing else to do" and at the same time it would say there
is data to be read in the FIFO, which is never possible.

This behaviour when using the continue bit caused random failures
in PIO mode read transfers of large sizes.

Instead of the continue bit, a GPIO can be used to control the
chipselect. This is handled by the SPI core.

After every transfer, polling for an ALLDONE interrupt is required.

BUG=chrome-os-partner:35348
TEST=SPI NAND test on BuB

Change-Id: Ie9d4b5aff0813bee909715dd8c13a34050415ab2
Signed-off-by: Sifan Naeem <sifan.naeem@imgtec.com>
Signed-off-by: Ezequiel Garcia <ezequiel.garcia@imgtec.com>
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/251980