CHROMIUM: spi: qup: Wait for QUP to complete
This patch changes the completion to occur when the QUP
controller specifies that all of the bytes have been
written/read.  This removes the timing issue that can
occur when the DMA completes the copy to the FIFO, but
the QUP has not sent out the bytes.  When this happens,
the QUP may not send out all the bytes in the FIFO before
it is placed back in the reset state.

BUG=chrome-os-partner:37905
TEST=Build and boot on Storm. Flashrom does not report
     "Verification" error on writing to NOR flash.

Change-Id: Ib41982c272daa3aa121ba773c3fa17c11ddf8a42
Signed-off-by: Andy Gross <agross@codeaurora.org>
Signed-off-by: Sarthak Kukreti <skukreti@codeaurora.org>
Reviewed-on: https://chromium-review.googlesource.com/262380
Reviewed-by: Bryan Freed <bfreed@chromium.org>
Commit-Queue: Bryan Freed <bfreed@chromium.org>
Tested-by: Bryan Freed <bfreed@chromium.org>