UPSTREAM: spi/rockchip: Round up clock rate divisor to err on the safe side
The Rockchip SPI driver currently calculates its clock rate divisor by
integer dividing the parent rate by the target rate, and then rounding
the result up to the next even number (since the divisor must be
even).

Clock rate divisors should always be rounded up, so that the resulting
frequency is lower or equal to the target. This is correctly done in the
second step here but not in the first, so we still have a risk of
exceeding the desired target frequency (e.g. setting spi-max-frequency
to 40000000 with a parent clock of 99000000 could lead to a divisor of
99000000 / 40000000 == 2 (which is even) that then results in an
effective frequency of 99000000 / 2 == 49500000 (potentially exceeding
the flash chip's specifications).

This patch changes the division to round up to fix this problem.

Signed-off-by: Julius Werner <jwerner@chromium.org>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 754ec43c0184aae48f5a8c917d8282449ab564a9)
Cherry-picked from broonie/spi.git/for-next with no conflicts.

BUG=chrome-os-partner:38339
TEST=Changed spi-max-frequency to 40000000 in the DTS, dumped the
divisor value during SPI reads and confirmed that it was 4 (for a
99000000 parent clock).

Change-Id: I93bb81c0f08bf9de1e2a8d440b1b4aa68e0393ab
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/262467
Reviewed-by: Douglas Anderson <dianders@chromium.org>