UPSTREAM: spi/rockchip: Add device tree property to configure Rx Sample Delay
We have found that we can sometimes see read failures on boards with
high-capacitance SPI lines. It seems that the controller samples the Rx
data line too early, and its register interface has an "Rx Sample Delay"
setting to fine-tune against this issue.

This patch adds a new optional device tree entry that can configure this
delay in terms of nanoseconds. The kernel will calculate the
best-fitting amount of parent clock ticks to program the controller with
based on that.

Signed-off-by: Julius Werner <jwerner@chromium.org>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 76b17e6e4923e4138ed9a4d480cba148ff239dc9)
Cherry-picked from broonie/spi.git/for-next with no conflicts.

BUG=chrome-os-partner:38339
TEST=Ran flashrom on Jerry with a 12ns delay, confirmed that it worked
at 49.5MHz and that the RSD value in CTRLR0 was '1' when booted with a
99MHz parent clock and '2' when booted with a 198MHz parent clock.

Change-Id: I85d7065dba08ba3c388f2381865b6a50cda744f9
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/262468
Reviewed-by: Douglas Anderson <dianders@chromium.org>