CHROMIUM: cpufreq: interactive: fix cpu load calculation
This uses the current frequency for calculating the cpu load instead of the
target load for a cpu. We still choose the same frequency when ignoring hispeed
freq, but we use the actual cpu load for choosing when to go to hispeed instead
of the artificially high one when we divide by that cpu's target freq.

BUG=chrome-os-partner:37673
TEST=run power_LoadTest on jerry
check we spend most of our time at the lower cpu frequencies

Change-Id: Id7e54614d12f883b8f40d87dd6c73101f2a0806e
Signed-off-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/263145
Reviewed-by: Douglas Anderson <dianders@chromium.org>