FROMLIST: mmc: dw_mmc: Increase cmd11 timeout to 500ms
The Designware databook claims that cmd11 should be finished in 2ms,
but my testing showed that not to be the case in some situations.
I've seen cmd11 timeouts of up to 130ms (!) during reboot tests.
Let's bump the timeout way up so that we're absolutely sure.  CMD11 is
only sent during card insertion, so this extra timeout shouldn't be
terrible.

BUG=chrome-os-partner:38719
TEST=reboot_test; add printouts to confirm

Fixes: 5c935165da79 ("mmc: dw_mmc: Add a timeout for sending CMD11")
Signed-off-by: Doug Anderson <dianders@chromium.org>
(am from https://patchwork.kernel.org/patch/6159461/)
Change-Id: I8266ede29edd20c991cf3a91814852ff0b241712
Reviewed-on: https://chromium-review.googlesource.com/263965
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>