CHROMIUM: drm: rockchip/dw_hdmi-rockchip: expand TMDS support list
This expands the table of pixel clocks based on needed / tested
resolutions.  Values come from the table in the user manual.

We've added these clocks:
31.50MHz, 49.50MHz, 50.00MHz, 71.00MHz, 73.25MHz, 75.00MHz, 78.75MHz,
85.50MHz, 88.75MHz, 115.50MHz, 119.00MHz, 135.00MHz, 136.75MHz,
154.00MHz, 162.00MHz

In order to be useful, some of these clocks will require to be
slightly adjusted in a later patch to match them up with PLL rates
that we can actually make.

BUG=chrome-os-partner:39069
TEST=jerry board, connect HDMI TV, change display resolutions,
     we can test part of clocks by the avaliable resolutions.

Change-Id: If5b19cd1c8107d00749ffca3272832981d5efb15
Signed-off-by: Yakir Yang <ykk@rock-chips.com>
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/265811