FROMLIST: ARM: rockchip: disable dapswjdp during suspend
Reset dapswjdp is controlled by JTAG_TRSTN, if the iomux of this pin is
not "jtag_trstn". the AP would think this pin is always high, so it can
not reset before resume. When system resume, but the dapswjdp is not in
a default state, it may Access some illegal address, it cause system
crash during resume.
Let's disable this jtag function by clear the dapdeviceen bit, it
prohibit the dapswjdp to access memory and registers. This bit would
be enable in MASKROM, so we need clear it in suspend everytime.

BUG=chrome-os-partner:38253
TEST=suspend_stress_test and dut could resume normally.

(am from https://patchwork.kernel.org/patch/6218781/)
Change-Id: I4b41418fe7bfe59ae5cb837c16d20445172a92bc
Signed-off-by: Chris Zhong <zyw@rock-chips.com>
Reviewed-on: https://chromium-review.googlesource.com/265602
Reviewed-by: Douglas Anderson <dianders@chromium.org>