CHROMIUM: rockchip: stop dmc devfreq on shutdown
This stops dmc clk transitions from happening when we shutdown. The way cpus are
paused in the clk set rate function can deadlock with how cpus are shutdown. If
the cpu that pauses the other cpus in dmc clk_set_rate is shutdown before it can
unpause the other cpus, we can fail to stop the secondary cpus.

BUG=chrome-os-partner:38951
TEST=run generic_RebootTest on speedy

Change-Id: I929d17c43a1afd9fa157a5d15e375478e8ed3ec2
Signed-off-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/265879
Reviewed-by: Douglas Anderson <dianders@chromium.org>