CHROMIUM: ARM: dts: Properly turn off pulls on emmc, sd, sdio
Due to an oversight we accidentally forgot to turn off internal pulls
for eMMC, SDMMC, and SDIO (WiFi).  We have external pulls on these
lines which have been calculated to be correct and adding the internal
pulls in addition isn't great.  Note also that we have the internal
pull enabled on the clock line where we really never want pulls.

Note that for the data / cmd lines, the internal pulls are unlikely to
matter _too_ much which is why this went unnoticed so long.  However,
the pulls have quite a bit of variability chip to chip, so it's
possible that they could matter more on some boards, especially with
higher speed cards.

From the datasheet, I see that pulls are ~34Kohm - 120Kohm, with
60Kohm being typical.

NOTE: this might cause some problems detecting SD cards on older
speedy boards (like speedy rev 2).  That board had the wrong pullup on
the CMD line, so having the extra internal pull actually _helped_.  SD
cards usually still work with the automatic SDMMC retries present in
the core and worst case a user could just remove / insert the card
again, but if we want to support SD on speed rev 2 we could either
pick something like CL:I611a65833051223a26dba31828a27d19661d4f61 or we
can fork its device tree and add a pullup just for that device.

BUG=chrome-os-partner:39745
TEST=Reboot tests; suspend stress tests; use "mem" to check pulls

Change-Id: I1e64b49e57831b219f10fcf3df9d2d791543e0e4
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/268799
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>