CHROMIUM: Input: atmel_mxt_ts - defer probe if reset GPIO is busy
We like to source different touch panels, but connect them to the same
regulators/GPIOs. If they are both probed at the same time one is loses
with -EBUSY. We can't sequence probes in any meaningful way since from
the driver core and i2c core these are 2 completely separate devices
residing at different addresses.

Let's convert -EBUSY from GPIO to -EPROBE_DEFER so that the loser gets
another chance (the drawback is that loser that does not actually have
corresponding device might be retrying probing for quite a few times -
but I do not see any easy way of fixing this).

This is not upstreamable IMO.

BUG=chrome-os-partner:39123
TEST=Build and boot on Minnie.

Change-Id: I797e017e488990dd18cf3bf50346fce45212784f
Signed-off-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/266225
Reviewed-by: Douglas Anderson <dianders@chromium.org>