FROMLIST: i2c: rk3x: Increase wait timeout to 1 second
While it's not sensible for an i2c command to _actually_ need more
than 200ms to complete, let's increase the timeout anyway.  Why?  It
turns out that if you've got a large number of printks going out to a
serial console, interrupts on a CPU can be disabled for hundreds of
milliseconds. That's not a great situation to be in to start with
(maybe we should put a cap in vprintk_emit()) but it's pretty annoying
to start seeing unexplained i2c timeouts.

A normal system shouldn't see i2c timeouts anyway, so increasing the
timeout should help people debugging without hurting other people
excessively.

BUG=chrome-os-partner:39321
TEST=Boot with lots of printks; no more timeouts

Signed-off-by: Doug Anderson <dianders@chromium.org>
(am from https://patchwork.kernel.org/patch/6307201/)
Change-Id: Ic6045c111eaa9f91703ce130c872c29754af0449
Reviewed-on: https://chromium-review.googlesource.com/268637
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Tested-by: Chris Zhong <zyw@rock-chips.com>