CHROMIUM: drivers: nss-ecm: Remove bad assertion check - race condition
Assertion was being checked after issuing NSS command.
However if NSS responds quicker - perhaps due to lock holdoff -
the assertion would race and trigger a bad condition.
The assertion is in the wrong place and is inappropriate.

BUG=chrome-os-partner:36182
TEST=build/boot on storm. Verified with the same stress test this
bad assert was reported and confirm it is not reproduceable

Change-Id: Id9b7273d8ee02a1d8cab9b809363a1d5e88a7bb8
Signed-off-by: Stephen Wang <wstephen@codeaurora.org>
Reviewed-on: https://chromium-review.googlesource.com/268825
Reviewed-by: Kevin Hayes <kevinhayes@chromium.org>
Reviewed-by: Eric Dumazet <edumazet@google.com>
Reviewed-by: Kevin Hayes <kevinhayes@google.com>
Tested-by: Kishan Kunduru <kkunduru@chromium.org>
Commit-Queue: Kishan Kunduru <kkunduru@chromium.org>