CHROMIUM: ARM: dts: Veyron: Pull up power button signal
On most devices (checked brain, mighty and jerry) the power button signal that
goes to the ap is only connected through a diode (so the diode can only make
the voltage be 0, otherwise it's floating). We really ought to keep it pulled
up instead otherwise it might not get high enough to trigger the logic.

BUG=chrome-os-partner:39899
TEST=Measure PWR_KEY_L, notice that it goes up to 3.3V instead of just 0.6V
TEST=Brain can now resume from suspend by pressing the power button.

Change-Id: I350438617de96b92fe72e9d4bdcbd9414567e1e1
Signed-off-by: Alexandru M Stan <amstan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/269823
Reviewed-by: Jerry Parson <jwp@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>