FROMLIST: watchdog: dw_wdt: keepalive the watchdog at write time
If you've got code that does this in a tight loop
  1. Open watchdog
  2. Send 'expect close'
  3. Close watchdog
...you'll eventually trigger a watchdog reset.  You can reproduce this
by using daisydog (1) and running:
  while true; do daisydog -c > /dev/null; done

The problem is that each time you write to the watchdog for 'expect
close' it moves the timer .5 seconds out.  The timer thus never fires
and never pats the watchdog for you.

1: http://git.chromium.org/gitweb/?p=chromiumos/third_party/daisydog.git

BUG=chrome-os-partner:39949
TEST=stop daisydog; while true; do daisydog -c >/dev/null; done

Signed-off-by: Doug Anderson <dianders@chromium.org>
(am from https://patchwork.kernel.org/patch/6360951/)
Change-Id: Ia215e9150d008c4d449300efb177f9dbf53cdfd2
Reviewed-on: https://chromium-review.googlesource.com/270002
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>