CHROMIUM: ARM: dts: Veyron: Use "disabled" status in lieu of "disable"
Per ePAPR 1.1 section 2.3.4, "disabled" is a valid status property but
"disable" is not.

Linux treats them as equivalent, so this shouldn't result in a behavioral
change.  __of_device_is_available() causes the device to be disabled if
the "status" property exists and the value matches anything except "okay"
or "ok".

BUG=none
TEST=boot test on minnie

Change-Id: I31562507d4caa7172ff6fe457b2eb48f8539f15a
Signed-off-by: Kevin Cernekee <cernekee@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/270248
Reviewed-by: Douglas Anderson <dianders@chromium.org>