CHROMIUM: ARM: dts: Veyron: Remove pinky
Pinky is a board we don't use anymore.
It hasn't been tested in a while, it might not boot anymore.

There's quite a couple of things that are defined in veyron*.dtsi where pinky
just inherets and all the other boards have to override
(ex: pwr_btn active low/high).  This is becoming a maintenance hassle for
a board that we don't use. So we should just remove it.

BUG=chrome-os-partner:39899
TEST=No more pinky

Change-Id: Ie3c8bd2e2a73c45aab5dd7110d4b880da82ebf65
Signed-off-by: Alexandru M Stan <amstan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/270273
Reviewed-by: Douglas Anderson <dianders@chromium.org>