Revert "CHROMIUM: netfilter: Add netfilter optional nf_ct_tcp_no_window_check"
This reverts commit 9c3420e6eeeb6aceaef0454b33c04dab5899f6ea.

The nf_ct_tcp_no_window_check was added in OpenWRT for performance
tuning,since QSDK is based on OpenWRT so we have this in QSDK kernel.
NSS is trying to act the same as the kernel stack so when this option
is on, we don't check tcp window in fastpath. Since this change only
gaining little performance improvement(5% in slowpath and 1% in
fastpath), this is decided to be removed from Chromium tree.

BUG=chrome-os-partner:40634
TEST=build/boot on whirlwind. Loaded on whirlwind and tested with NSS turned on and off.

Change-Id: I07afb2f044375331d9d793938bf7281e511972ae
Signed-off-by: Stephen Wang <wstephen@codeaurora.org>
Reviewed-on: https://chromium-review.googlesource.com/271601
Reviewed-by: Kishan Kunduru <kkunduru@chromium.org>
Commit-Queue: Kishan Kunduru <kkunduru@chromium.org>
Tested-by: Kishan Kunduru <kkunduru@chromium.org>